Revert "New Changes"
This reverts commit 8bc6393c9a038e35b9847ae32f3a968b53f60a23.
