+++ nbproject/private/private.xml
@@ -3,11 +3,9 @@
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
-            <file>file:/C:/Users/Inspiron/Documents/GitHub/CPT-163-27-F2015-Assignment-7-Array-DiceTower/src/dnddiceroller/Die.java</file>
             <file>file:/C:/Users/Inspiron/Documents/GitHub/CPT-163-27-F2015-Assignment-7-Array-DiceTower/src/dnddiceroller/DnDDiceRoller.java</file>
-            <file>file:/C:/Users/Inspiron/Documents/GitHub/CPT-163-27-F2015-Assignment-7-Array-DiceTower/src/dnddiceroller/NumberedDie.java</file>
-            <file>file:/C:/Users/Inspiron/Documents/GitHub/CPT-163-27-F2015-Assignment-7-Array-DiceTower/src/dnddiceroller/LoadedDice.java</file>
-            <file>file:/C:/Users/Inspiron/Documents/GitHub/CPT-163-27-F2015-Assignment-7-Array-DiceTower/src/dnddiceroller/FudgeDice.java</file>
+            <file>file:/C:/Users/Inspiron/Documents/GitHub/CPT-163-27-F2015-Assignment-7-Array-DiceTower/src/dnddiceroller/Die.java</file>
+            <file>file:/C:/Users/Inspiron/Documents/GitHub/CPT-163-27-F2015-Assignment-7-Array-DiceTower/src/dnddiceroller/DiceTower.java</file>
         </group>
     </open-files>
 </project-private>
