+++ src/dnddiceroller/DnDDiceRoller.java
@@ -35,7 +35,7 @@ public class DnDDiceRoller {
   }
   
   public static int test_oneD6(){
-	NumberedDie die = new NumberedDie(6);
+	Die die = new Die(6);
 	die.roll();
 	int dieValue = die.value();
 	if(dieValue < 1 || dieValue > 6){
@@ -46,9 +46,9 @@ public class DnDDiceRoller {
   }
   
   public static int test_diceTowerWithTwoD6(){
-	List<NumberedDie> dice = new ArrayList();
-	dice.add(new NumberedDie(6));
-	dice.add(new NumberedDie(6));
+	List<Die> dice = new ArrayList();
+	dice.add(new Die(6));
+	dice.add(new Die(6));
 	DiceTower tower = new DiceTower(dice);
 	tower.dropDice();
 	int trayValue = tower.trayValue();
