--- src/dnddiceroller/FudgeDice.java
@@ -1,24 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package dnddiceroller;
-
-/**
- *
- * @author Inspiron
- */
-public class FudgeDice implements Die{
-
-    @Override
-    public void roll() {
-        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
-    }
-
-    @Override
-    public int value() {
-        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
-    }
-    
-}
